Move isDateTimeOrTemporal
